Extract binding backfill into nested Backfill service
