+++ test/models/apple/show_feed_binding/backfill_test.rb
@@ -1,0 +1,108 @@
+require "test_helper"
+
+module Apple
+  describe ShowFeedBinding::Backfill do
+    describe ".backfill!" do
+      it "creates a binding and sets the config from a public feed sync log" do
+        config = create_config_with_legacy_show_id(sync_log_show_id: "show-from-sync", private_show_id: "show-from-feed")
+
+        report = ShowFeedBinding::Backfill.backfill!
+
+        binding = config.public_feed.apple_show_feed_binding
+        assert_equal 1, report[:created]
+        assert_equal 1, report[:linked]
+        assert_equal binding, config.reload.show_feed_binding
+        assert_equal config.public_feed, binding.feed
+        assert_equal config.key, binding.apple_key
+        assert_equal "show-from-sync", binding.apple_show_id
+      end
+
+      it "falls back to the private feed apple show id" do
+        config = create_config_with_legacy_show_id(private_show_id: "show-from-feed")
+
+        ShowFeedBinding::Backfill.backfill!
+
+        assert_equal "show-from-feed", config.reload.show_feed_binding.apple_show_id
+      end
+
+      it "skips and reports configs missing a key or show id" do
+        missing_key = create_config_with_legacy_show_id(private_show_id: "show-without-key", key: nil)
+        missing_show_id = create_config_with_legacy_show_id
+
+        report = ShowFeedBinding::Backfill.backfill!
+
+        assert_equal 2, report[:skipped]
+        assert_equal [missing_key.id, missing_show_id.id].sort, report[:skipped_configs].map { |row| row[:config_id] }.sort
+        assert_includes report[:skipped_configs].map { |row| row[:reason] }, "missing key"
+        assert_includes report[:skipped_configs].map { |row| row[:reason] }, "missing show id"
+      end
+
+      it "is idempotent" do
+        create_config_with_legacy_show_id(sync_log_show_id: "show-from-sync")
+        ShowFeedBinding::Backfill.backfill!
+
+        assert_no_difference "ShowFeedBinding.count" do
+          report = ShowFeedBinding::Backfill.backfill!
+
+          assert_equal 0, report[:changed]
+          assert_equal 0, report[:created]
+          assert_equal 0, report[:updated]
+          assert_equal 0, report[:linked]
+        end
+      end
+
+      it "can dry run without writing" do
+        config = create_config_with_legacy_show_id(sync_log_show_id: "show-from-sync")
+
+        assert_no_difference "ShowFeedBinding.count" do
+          report = ShowFeedBinding::Backfill.backfill!(dry_run: true)
+
+          assert_equal 1, report[:created]
+          assert_equal 1, report[:linked]
+          assert_equal "would_create", report[:actions].first[:action]
+        end
+        assert_nil config.reload.show_feed_binding
+      end
+    end
+
+    describe ".verify_routing_equivalence!" do
+      it "reports zero mismatches after backfill" do
+        create_config_with_legacy_show_id(sync_log_show_id: "show-from-sync")
+        ShowFeedBinding::Backfill.backfill!
+
+        report = ShowFeedBinding::Backfill.verify_routing_equivalence!
+
+        assert_empty report[:mismatches]
+      end
+
+      it "reports a seeded mismatch" do
+        config = create_config_with_legacy_show_id(sync_log_show_id: "show-from-sync")
+        ShowFeedBinding::Backfill.backfill!
+        config.reload.show_feed_binding.update!(apple_show_id: "wrong-show")
+
+        report = ShowFeedBinding::Backfill.verify_routing_equivalence!
+
+        assert_equal 1, report[:mismatches].length
+        assert_equal config.id, report[:mismatches].first[:config_id]
+        assert_equal "apple_show_id", report[:mismatches].first[:mismatches].first[:field]
+      end
+    end
+
+    def create_config_with_legacy_show_id(sync_log_show_id: nil, private_show_id: nil, key: create(:apple_key))
+      podcast = create(:podcast)
+      private_feed = create(:private_feed, podcast: podcast, apple_show_id: private_show_id)
+      config = create(:apple_config, feed: private_feed, key: key)
+
+      if sync_log_show_id
+        SyncLog.log!(
+          integration: :apple,
+          feeder_type: :feeds,
+          feeder_id: podcast.public_feed.id,
+          external_id: sync_log_show_id
+        )
+      end
+
+      config
+    end
+  end
+end
